+++ app/context/index.ts
@@ -1,5 +1,7 @@
 export * from './PlannerContext';
 export * from './PlannerProvider';
+export * from './ProjectsContext';
+export * from './ProjectsProvider';
 export * from './ViewModeContext';
 export * from './ProjectsViewContext';
 export * from './CalendarViewContext';